+++ NEWS.md
@@ -4,8 +4,10 @@
 
 - add interactive version of `ggbeeswarm::geom_quasirandom`, see
 `geom_quasirandom_interactive()`.
-- add `default_on` parameter to `opts_zoom()` to automatically activate pan/zoom 
+- add `default_on` parameter to `opts_zoom()` to automatically activate pan/zoom
 mode when plot is rendered.
+- improve `opts_toolbar()` documentation and examples to better demonstrate the `hidden`
+parameter for customizing which toolbar buttons are displayed.
 
 ## Issues